MoveInstructionGenerator removed
