--- Assets/Scripts/Server/Instructions/MoveInstructionGenerator.asset
@@ -1,15 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: f339ad68a5b8f6749896278aea7f3220, type: 3}
-  m_Name: MoveInstructionGenerator
-  m_EditorClassIdentifier: 
-  f: 0
