--- Assets/Scripts/Server/Instructions/MoveInstructionGenerator.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 17d01433fe082e543b9737d3a124576f
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
